pyproject.toml
[project]
name = "crypto-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.21.0",
    "loguru>=0.7.3",
    "peewee>=3.18.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
